I've Cracked It
I was calling the filter function from my old chart which parses the "value" out of the data instead of the "members" which is why every chart was zero-ing out. It took me way too long to spot this, but now I've got it and I can work on multiple simultaneous selections. Graying out bars out of focus in the selected chart is functional. Bars reorient after focus shifts to their chart.