Never been so happy to see my project break
I have (an incorrect version of) multi-dimensional filtering. It currently zeroes out in instances where I don't think it should, but I'm making progress! I had to make adjustments to clickhandler and the fourBars directive, and create an array in rootScope. The clickhandler now takes all four dimensions and filters all of them based on all of the filter values that have been passed in. Those values are stored in an array (dims) in rootscope.