Multi-Dimensional Filtering
All dimensions to filter on and the appropriate filters are stored in objects, pushed to an array in $rootScope, and passed to the click function. This allows which filters have been used to be persisted. I stop it at 4 since there are 4 charts. I have some more work to do trying to remove hardcoded values from the controller in fourBars, but it's *mostly* good to go.