Commit After Failed Attempt at Data Extraction
I took a stab at using the controller to extract the different categories of data from the data array so I don't need the hardcoded 'names' array in fourBars, but it ended up breaking most everything so I reverted. Attempt two will be to try a similar approach in the factory.