Fully Scalable
No longer hardcoded:
-Array of names
-Dimensions
-Sets of values
All functions that used these values when they were hardcoded now accept arrays that contain what used to the hardcoded values. The lynchpin function is extractNames in factory.js which extracts the category names from the JSON data. The filename is still hardcoded but there's no reason to change that right now. In theory this can handle sets with many more parameters. The only place a new category would need to be added is in BarChartController.js so as to extend the loop in bar2.html and create a new svg element.