Style Changes and Groundwork for Next Step
The charts are smaller and (at least on my computer) two are showing up in a row. I might need to make alterations so that remains constant. Buttons for the next step are in place. An empty function to add tooltips has been created. A function to filter out the funds values has been added, as have the resultant arrays, but the arrays are not used yet.