Header Updated to Include Filter Values and Working Reset
The reset button sets every chart to its default position. The header includes which values are currently being used by the filter. The charts are in rows of two. There are tooltips for every bar. Next step is switching between members and funds without filters, then making sure it works with filters.