Mostly Functional
Weird stuff happens when filtering on three dimensions, but it's mostly correct! I'll tune it up in the new year, but this is as good a place to stop as any for now. There is also an unresolved bugs with tooltips (temporary fix causes tooltips to update incorrectly for the third bar in charts 1-3) that needs to be fixed