Functional Chart
I'll call this a finished prototype. I'm trying to parameterize the coloring of the bars so that the Actual Time bars will have a color based on another value I'm going to add to the objects. So tasks in progress, complete, and over time are shown in different colors in the actual row. I could probably also stand to add more detail to my comments.